extract db_connection fixture to conftext.py to use across modules
